Updated index.html
I figured out how to show the image on the app! Earlier it was only showing the alt text, and when I would load the image to a new tab, the image could not be found. I was having issues with the folder path and making sure the image was easy to locate. The image should be visibale and resizeable now!